Add files via upload
We are a passionate group of fitness enthusiasts and web developers who came together to create an exciting fitness website. Our project is a collaborative effort that combines HTML, CSS, and JavaScript to provide an interactive and engaging platform for all fitness enthusiasts. With a shared commitment to health and wellness, we've poured our knowledge and skills into this project to help individuals achieve their fitness goals. Explore our code, see the results of our hard work, and join us in promoting a healthier and fitter world through the power of technology. Let's get fit together, one line of code at a time! 💪💻 #FitTech #WebDev #FitnessJourney